+++ 10.0.10240.16384/ntoskrnl/Functions/__jump_unwind.c
@@ -5,5 +5,5 @@ void __fastcall __noreturn _jump_unwind(int a1, int a2)
   char v2; // [sp+10h] [bp-1A8h] BYREF
 
   while ( 1 )
-    a1 = RtlUnwindEx(a1, a2, 0, 0, &v2, 0);
+    a1 = RtlUnwindEx(a1, a2, 0, 0, (int)&v2, 0);
 }
